Move VIconWrapper and Vuetify icons helper to Thumbnail dir
This is part of migrating Studio to KIcon.

Thumbnail.vue component is the only remaning component using
VIconWrapper. It can't be migrated to KIcon yet since
KIcon needs to be updated at first with some new features.

Moving the files here ensures that VIconWrapper won't be used
from anywhere else in the Studio.
