+++ contentcuration/contentcuration/frontend/shared/views/files/Thumbnail/VIconWrapper.vue
