+++ contentcuration/contentcuration/frontend/shared/views/files/Thumbnail/icons.js
